Following classes were moved to the Kx sub-project - KxButton - KxComboBox - KxBitmapComboBox - KxLabel - KxTextBox - KxStyledTextBox - KxProgressBar - KxDialog - KxStdDialog - KxSimpleStdDialog - KxCredentialsDialog - KxFileBrowseDialog - KxTextBoxDialog - KxTaskDialog (WIP) - KxProgressDialog
Removed:
- KxDualProgressDialog
- KxDualInputDialog

+++ Kx/UI/Controls/StatusBar.cpp
@@ -18,7 +18,7 @@ namespace KxFramework::UI
 			{
 				std::vector<int> newWidths;
 				newWidths.resize(count);
-				for (size_t i = 0; i < count; i++)
+				for (size_t i = 0; i < static_cast<size_t>(count); i++)
 				{
 					newWidths[i] = GetStatusWidth(i);
 				}
